FuturesBot/requirements.txt
websocket-client
jsonpickle 
numpy 
datetime 
requests 
python-binance 
binance.py
binance-futures
pyTelegramBotAPI